move host code out of NetMan and into NS
