+++ Assets/Mirror/Runtime/NetworkManager.cs
@@ -1,7 +1,4 @@
-using System;
-using System.Threading.Tasks;
 using UnityEngine;
-using UnityEngine.Events;
 
 namespace Mirror
 {
@@ -11,10 +8,8 @@ namespace Mirror
     [RequireComponent(typeof(NetworkServer))]
     [RequireComponent(typeof(NetworkClient))]
     [DisallowMultipleComponent]
-    public class NetworkManager : MonoBehaviour, INetworkManager
+    public class NetworkManager : MonoBehaviour
     {
-        static readonly ILogger logger = LogFactory.GetLogger<NetworkManager>();
-
         public NetworkServer server;
         public NetworkClient client;
 
@@ -23,52 +18,5 @@ namespace Mirror
         /// <para>This is set True in StartServer / StartClient, and set False in StopServer / StopClient</para>
         /// </summary>
         public bool IsNetworkActive => server.Active || client.Active;
-
-        /// <summary>
-        /// This is invoked when a host is started.
-        /// <para>StartHost has multiple signatures, but they all cause this hook to be called.</para>
-        /// </summary>
-        public UnityEvent OnStartHost = new UnityEvent();
-
-        /// <summary>
-        /// This is called when a host is stopped.
-        /// </summary>
-        public UnityEvent OnStopHost = new UnityEvent();
-
-        /// <summary>
-        /// This starts a network "host" - a server and client in the same application.
-        /// <para>The client returned from StartHost() is a special "local" client that communicates to the in-process server using a message queue instead of the real network. But in almost all other cases, it can be treated as a normal client.</para>
-        /// </summary>
-        public async Task StartHost()
-        {
-            if (!server)
-                throw new InvalidOperationException("NetworkServer not assigned. Unable to StartHost()");
-            if (!client)
-                throw new InvalidOperationException("NetworkClient not assigned. Unable to StartHost()");
-
-            // start listening to network connections
-            await server.ListenAsync();
-
-            client.ConnectHost(server);
-
-            // call OnStartHost AFTER SetupServer. this way we can use
-            // NetworkServer.Spawn etc. in there too. just like OnStartServer
-            // is called after the server is actually properly started.
-            OnStartHost.Invoke();
-
-            logger.Log("NetworkManager StartHost");
-        }
-
-        /// <summary>
-        /// This stops both the client and the server that the manager is using.
-        /// </summary>
-        public void StopHost()
-        {
-            OnStopHost.Invoke();
-            if(client)
-                client.Disconnect();
-            if(server)
-                server.Disconnect();
-        }
     }
 }
