+++ Assets/Tests/Performance/Runtime/BenchmarkPerformanceLight.cs
@@ -36,7 +36,7 @@ namespace Mirror.Tests.Performance.Runtime
                 yield break;
             }
 
-            System.Threading.Tasks.Task task = benchmarker.StartHost();
+            System.Threading.Tasks.Task task = benchmarker.server.StartHost(benchmarker.client);
 
             while (!task.IsCompleted)
                 yield return null;
@@ -47,7 +47,7 @@ namespace Mirror.Tests.Performance.Runtime
         public IEnumerator TearDown()
         {
             // shutdown
-            benchmarker.StopHost();
+            benchmarker.server.StopHost();
             yield return null;
 
             // unload scene
